Final-Year-Project-main/Server/requirements.txt
Flask
pandas
scikit-learn
Flask-Cors
python-dotenv
google-generativeai==0.7.2
requests